+++ src/swaps/mod.rs
@@ -5,6 +5,7 @@ pub mod fees;
 pub mod kaleido;
 pub mod liquid;
 pub mod magic_routing;
+pub mod rgb;
 #[cfg(feature = "ws")]
 mod status_stream;
 mod wrappers;